+++ metadata_annotator/annotate.py
@@ -374,7 +374,7 @@ def get_grid_start_index(
     if subset_index_reference:
         return get_start_index_from_row_col_variable(datatree, subset_index_reference)
     if (
-        dim_data_array.attrs.get('corner_points_offsets', None)
+        dim_data_array.attrs.get('corner_point_offsets', None)
         == 'history_subset_index_ranges'
     ):
         return get_start_index_from_history(