+++ metadata_annotator/annotate.py
@@ -373,10 +373,7 @@ def get_grid_start_index(
     subset_index_reference = dim_data_array.attrs.get('subset_index_reference', None)
     if subset_index_reference:
         return get_start_index_from_row_col_variable(datatree, subset_index_reference)
-    if (
-        dim_data_array.attrs.get('corner_point_offsets', None)
-        == 'history_subset_index_ranges'
-    ):
+    if dimension_index_map:
         return get_start_index_from_history(
             dimension_index_map, dimension_variable_path
         )