+++ CHANGELOG.md
@@ -7,11 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-04-24
 
-### Changed
-
 - Update Harmony Metadata Annotator to rename pseudo dimension variables
   that get created in collections that do not have dimension variables.
   Rename the dimension variables with the names provided in the earthdata-varinfo configuration file.
+
 
 ## [unreleased] - 2025-04-22
 
@@ -21,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   the global grid mapping variables resolution specific, add a `master_geotransform`
   attribute to all grid mapping variables, add missing dimension variables, and create a dimensions
   attribute on their associated variables.
+
 
 ## [v1.0.0] - 2025-03-28