+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased] - 2025-05-07
+
+- Use the index ranges from the history attribute to calculate the grid start index. The dimension 
+  scales are then determined based on the projection attributes provided in the associated grid 
+  mapping variable. 
+
+
+## [unreleased] - 2025-05-06
+
+- Update the Harmony Metadata Annotator to update spatial dimension variables with a dimension
+  scale that is computed based off the associated grid mapping variable's `master_geotransform` and
+  the determined grid start index. Currently, the grid start index can be determined from the
+  dimension's associated row/col index variable.
+
+
 ## [unreleased] - 2025-04-24
 
 - Update Harmony Metadata Annotator to rename pseudo dimension variables